For #11783, rename package and type name.
