+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/java/org/apache/shardingsphere/distsql/parser/api/DistSQLStatementParserEngine.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.distsql.parser.api;
 import org.antlr.v4.runtime.misc.ParseCancellationException;
 import org.apache.shardingsphere.distsql.parser.core.advanced.AdvancedDistSQLStatementParserEngine;
 import org.apache.shardingsphere.distsql.parser.core.common.CommonDistSQLStatementParserEngine;
-import org.apache.shardingsphere.distsql.parser.core.rule.FeaturedDistSQLStatementParserEngine;
+import org.apache.shardingsphere.distsql.parser.core.featured.FeaturedDistSQLStatementParserEngine;
 import org.apache.shardingsphere.sql.parser.exception.SQLParsingException;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
