+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-distsql/shardingsphere-db-discovery-distsql-parser/src/main/java/org/apache/shardingsphere/dbdiscovery/distsql/parser/facade/DatabaseDiscoveryDistSQLStatementParserFacade.java
@@ -46,7 +46,7 @@ public final class DatabaseDiscoveryDistSQLStatementParserFacade implements Feat
     }
     
     @Override
-    public String getRuleType() {
+    public String getFeatureType() {
         return "database-discovery";
     }
 }
