+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-distsql/shardingsphere-encrypt-distsql-parser/src/main/java/org/apache/shardingsphere/encrypt/distsql/parser/facade/EncryptDistSQLStatementParserFacade.java
@@ -46,7 +46,7 @@ public final class EncryptDistSQLStatementParserFacade implements FeaturedDistSQ
     }
     
     @Override
-    public String getRuleType() {
+    public String getFeatureType() {
         return "encrypt";
     }
 }
