+++ shardingsphere-features/shardingsphere-readwrite-splitting/shardingsphere-readwrite-splitting-distsql/shardingsphere-readwrite-splitting-distsql-parser/src/main/java/org/apache/shardingsphere/readwritesplitting/distsql/parser/facade/ReadwriteSplittingDistSQLStatementParserFacade.java
@@ -46,7 +46,7 @@ public final class ReadwriteSplittingDistSQLStatementParserFacade implements Fea
     }
     
     @Override
-    public String getRuleType() {
+    public String getFeatureType() {
         return "readwrite-splitting";
     }
 }
