+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/shardingsphere-sharding-distsql-parser/src/main/java/org/apache/shardingsphere/sharding/distsql/parser/facade/ShardingDistSQLStatementParserFacade.java
@@ -46,7 +46,7 @@ public final class ShardingDistSQLStatementParserFacade implements FeaturedDistS
     }
     
     @Override
-    public String getRuleType() {
+    public String getFeatureType() {
         return "sharding";
     }
 }
