+++ shardingsphere-scaling/shardingsphere-scaling-distsql/shardingsphere-scaling-distsql-parser/src/main/java/org/apache/shardingsphere/scaling/distsql/parser/facade/ScalingDistSQLStatementParserFacade.java
@@ -46,7 +46,7 @@ public final class ScalingDistSQLStatementParserFacade implements FeaturedDistSQ
     }
     
     @Override
-    public String getRuleType() {
+    public String getFeatureType() {
         return "scaling";
     }
 }
